[INTERNAL] Make the plugin buildable on our Jenkins system
Asked Eclipse to create an OSGI Bundle Manifest, rename the test to
end with *Test to be picked up by the test task. The location of the
test was changed to help ant pick the right package name for junit. I
couldn't easily figure out how ant derives the package name. On first
sight this setup appeared to be like .dpp/ but some details are not the
same.

Change-Id: I885709ec3f4a38cf1f2654c6f4145df35e7196da
Reviewed-on: http://saros-build.imp.fu-berlin.de/gerrit/103
Tested-by: Jenkins CI
Reviewed-by: Holger Freyther <holger@freyther.de>